fix lots of minor problems
1) Use logging facility with timestamp and pid for most messages. Makes
debugging easier.

2) Timestamp on Kiwi URL was not unique if multiple connections made
during the same second. Salted with pid and instance index in the case
of a multi-connection configuration. This was causing the Kiwi to
reject what it considered to be a duplicate connection attempt due to
the timestamp non-uniqueness. For various reasons the Kiwi can’t
consider the connection port number which of course will be different
for multiple connections from the same host.

3) A multi-connection run without differentiating the connections with
the —station option caused all the connection processes to write to the
same filename. Detect this case and add connection index to filename.

4) Sometimes in the .wav header update code the fp.tell() was returning
zero for the file size. This is bad because the code attempts to write
“filesize - 8” into the header as an unsigned value via struct.pack()
which throws an exception. This exception is caught and ignored and so
isn’t an explanation for the current crashing/disconnect problem.

5) Made the default launch delay zero seconds instead of one since the
new timestamp allows for this.